--- android/src/main/kotlin/com/mytiki/sdk/capture/receipt/capacitor/rsp/RspEmailAccount.kt
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) TIKI Inc.
- * MIT license. See LICENSE file in root directory.
- */
-
-package com.mytiki.sdk.capture.receipt.capacitor.rsp
-
-import org.json.JSONObject
-
-class RspEmailAccount(
-    private val username: String,
-    private val provider: String,
-    private val verified: Boolean
-) : Rsp {
-    override fun toJson(): JSONObject =
-        JSONObject()
-            .put("username", username)
-            .put("provider", provider)
-            .put("verified", verified)
-}
